se fosse no gmail...

+++ Projeto/src/Nathalya/Aluno.java
@@ -46,12 +46,13 @@ public class Aluno implements Comparable<Aluno> {
 		return nome;
 	}
 
+
 	/**
 	 * Retorna a nota do aluno 
 	 * @return nota do aluno 
 	 */
 	
-	public int getNota() {
+	public double getNota() {
 		return nota;
 	}
